requirements.txt
boto3
pandas
pymongo
python-dotenv